backend/requirements.txt
flask==3.0.3
gunicorn==22.0.0
mysql-connector-python==9.0.0
tzdata>=2024.1

